pyproject.toml
[tool.poetry]
name = "serraniel-psacard"
version = "0.1.0"
description = ""
authors = ["Der-Eddy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
progressbar2 = "^4.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
